Fixes #4655 -- Move localize setup to resolve import loop
Localize has an import loop, because submodules are importing
from their root. To resolve this `msg`, `installed` and `_installMsgImplementation`
have moved to a new submodule.

Imports to `msg` stay intact, through an explicit export.
