+++ .changeset/chilly-melons-laugh.md
@@ -1,0 +1,5 @@
+---
+"@lit/localize": patch
+---
+
+Move localize setup to resolve import loop
